Additional puzzle work
Finished tiling the puzzle
Added an instance layer to rm_Puzzle for deathzones
Added spawn object to the puzzle
Changed viewport size for the puzzle so more of it is visible at one time. I think the puzzle is too hard to navigate because the player can't see where they are going. Happy for any feedback/thoughts on that choice
Added checkpoint layer to lvl1 and added in death zones and respawn
Changed sprite names for consistency in naming convenion (spr_enemy_stri[4 > s_enemy_strip4 and spr_level1Tiles > s_level1Tiles)