Puzzle edits
New o_wall object for the puzzle. It's a placeholder
Removed warp from the start of the puzzle level which was breaking the game
Retiled the puzzle
Created room for level 3 (rm_lvl3)
Added step event to o_weakwall. The object deactivates when the player collides with it and dashes
Edited creation code in level warp. It now references to x and y pos of the o_tutorial_spawn
Added chasing bug to tutorial and lvl1
Bug moves automatically (it follows the camera target) and restarts the room if it collides with the player object
Created a new instance layer for the player and enemy. Lets them appear, whilst the checkpoints etc. are made invisible